INS-1926: add node profile to context

+++ network/servicenetwork/servicenetwork.go
@@ -283,7 +283,7 @@ func (n *ServiceNetwork) HandlePulse(ctx context.Context, newPulse insolar.Pulse
 	n.lock.Lock()
 	defer n.lock.Unlock()
 
-	ctx = utils.NewPulseContext(ctx, uint64(newPulse.PulseNumber))
+	ctx = utils.NewPulseContext(ctx, uint32(newPulse.PulseNumber))
 	logger.Infof("Got new pulse number: %d", newPulse.PulseNumber)
 	ctx, span := instracer.StartSpan(ctx, "ServiceNetwork.Handlepulse")
 	span.AddAttributes(
